only set up certs if MINIO_SELF_CERT is 1

+++ lib/minio/minio/setup.sh
@@ -1,13 +1,16 @@
 #!/usr/bin/env bash
 
 set -Euo pipefail
-rm /root/.minio/certs
-mkdir /root/.minio/certs
-cp /run/secrets/private.key /root/.minio/certs
-cp /run/secrets/public.crt /root/.minio/certs
 
-mkdir /root/.minio/certs/$CANDIG_DOMAIN
-cp /run/secrets/private.key /root/.minio/certs/$CANDIG_DOMAIN/
-cp /run/secrets/public.crt /root/.minio/certs/$CANDIG_DOMAIN/
+if [ $MINIO_SELF_CERT -eq 1 ]; then
+  rm /root/.minio/certs
+  mkdir /root/.minio/certs
+  cp /run/secrets/private.key /root/.minio/certs
+  cp /run/secrets/public.crt /root/.minio/certs
+
+  mkdir /root/.minio/certs/$CANDIG_DOMAIN
+  cp /run/secrets/private.key /root/.minio/certs/$CANDIG_DOMAIN/
+  cp /run/secrets/public.crt /root/.minio/certs/$CANDIG_DOMAIN/
+fi
 
 top -b
